feat: add LU decomposition algorithm
Also replaced long long for long double in all files.

+++ inc/Functions.hpp
@@ -6,8 +6,8 @@
 #include "Matrix.hpp"
 #include "TerminalFunctions.hpp"
 
-std::vector<std::vector<long long>> *getTable(); // Return a matrix given in the command line
+std::vector<std::vector<long double>> *getTable(); // Return a matrix given in the command line
 Matrix getMatrixSaved(std::vector<Matrix *>);
-long long getNumber();
+long double getNumber();
 
 #endif
